new cmdsys initial commit
